+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/ConfigToRequestTransformer.java
@@ -95,6 +95,10 @@ final class ConfigToRequestTransformer {
     Option.of(config.getFilteringConfig()).forEach(moduleConfig::filteringModuleConfig);
     Option.of(config.getMaskingConfig()).forEach(moduleConfig::maskingModuleConfig);
     Option.of(config.getGroundingConfig()).forEach(moduleConfig::groundingModuleConfig);
+    Option.of(config.getOutputTranslationConfig())
+        .forEach(moduleConfig::outputTranslationModuleConfig);
+    Option.of(config.getInputTranslationConfig())
+        .forEach(moduleConfig::inputTranslationModuleConfig);
 
     return moduleConfig;
   }