+++ sample-code/spring-app/src/test/java/com/sap/ai/sdk/app/controllers/OrchestrationTest.java
@@ -19,6 +19,7 @@ import com.sap.ai.sdk.orchestration.TemplateConfig;
 import com.sap.ai.sdk.orchestration.TextItem;
 import com.sap.ai.sdk.orchestration.model.CompletionPostResponse;
 import com.sap.ai.sdk.orchestration.model.DPIEntities;
+import com.sap.ai.sdk.orchestration.model.GenericModuleResult;
 import com.sap.ai.sdk.orchestration.model.LLMChoice;
 import com.sap.ai.sdk.orchestration.model.LLMModuleResultSynchronous;
 import java.io.IOException;
@@ -32,16 +33,17 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 
 @Slf4j
 class OrchestrationTest {
-  OrchestrationService service;
   private final OrchestrationClient client = new OrchestrationClient();
   private final OrchestrationModuleConfig config =
       new OrchestrationModuleConfig().withLlmConfig(GEMINI_1_5_FLASH.withParam(TEMPERATURE, 0.0));
+  OrchestrationService service;
 
   @BeforeEach
   void setUp() {
@@ -50,7 +52,7 @@ class OrchestrationTest {
 
   @Test
   void testCompletion() {
-    final var result = service.completion("HelloWorld!");
+    val result = service.completion("HelloWorld!");
 
     assertThat(result).isNotNull();
     assertThat(result.getContent()).isNotEmpty();
@@ -58,10 +60,10 @@ class OrchestrationTest {
 
   @Test
   void testStreamChatCompletion() {
-    final var prompt = new OrchestrationPrompt("Who is the prettiest?");
-    final var stream = new OrchestrationClient().streamChatCompletion(prompt, service.getConfig());
+    val prompt = new OrchestrationPrompt("Who is the prettiest?");
+    val stream = new OrchestrationClient().streamChatCompletion(prompt, service.getConfig());
 
-    final var filledDeltaCount = new AtomicInteger(0);
+    val filledDeltaCount = new AtomicInteger(0);
     stream
         // foreach consumes all elements, closing the stream at the end
         .forEach(
@@ -80,10 +82,10 @@ class OrchestrationTest {
   @Test
   void testTemplate() {
     assertThat(service.getConfig().getLlmConfig()).isNotNull();
-    final var modelName = service.getConfig().getLlmConfig().getModelName();
+    val modelName = service.getConfig().getLlmConfig().getModelName();
 
-    final var result = service.template("German");
-    final var response = result.getOriginalResponse();
+    val result = service.template("German");
+    val response = result.getOriginalResponse();
 
     assertThat(response.getRequestId()).isNotEmpty();
     assertThat(((TextItem) result.getAllMessages().get(0).content().items().get(0)).text())
@@ -124,7 +126,7 @@ class OrchestrationTest {
   void testMessagesHistory() {
     CompletionPostResponse result =
         service.messagesHistory("What is the capital of France?").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
@@ -274,12 +276,12 @@ class OrchestrationTest {
 
   @Test
   void testImageInput() {
-    final var result =
+    val result =
         service
             .imageInput(
                 "https://upload.wikimedia.org/wikipedia/commons/thumb/5/59/SAP_2011_logo.svg/440px-SAP_2011_logo.svg.png")
             .getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
@@ -297,33 +299,33 @@ class OrchestrationTest {
     } catch (Exception e) {
       System.out.println("Error fetching or reading the image from URL: " + e.getMessage());
     }
-    final var result = service.imageInput(dataUrl).getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val result = service.imageInput(dataUrl).getOriginalResponse();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testMultiStringInput() {
-    final var result =
+    val result =
         service
             .multiStringInput(
                 List.of("What is the capital of France?", "What is Chess about?", "What is 2+2?"))
             .getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testResponseFormatJsonSchema() {
-    final var result = service.responseFormatJsonSchema("apple").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val result = service.responseFormatJsonSchema("apple").getOriginalResponse();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testResponseFormatJsonObject() {
-    final var result = service.responseFormatJsonObject("apple").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val result = service.responseFormatJsonObject("apple").getOriginalResponse();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
     assertThat(choices.get(0).getMessage().getContent()).contains("\"language\":");
     assertThat(choices.get(0).getMessage().getContent()).contains("\"translation\":");
@@ -331,46 +333,45 @@ class OrchestrationTest {
 
   @Test
   void testResponseFormatText() {
-    final var result = service.responseFormatText("apple").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val result = service.responseFormatText("apple").getOriginalResponse();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testTemplateFromPromptRegistryById() {
-    final var result =
-        service.templateFromPromptRegistryById("Cloud ERP systems").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val result = service.templateFromPromptRegistryById("Cloud ERP systems").getOriginalResponse();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testTemplateFromPromptRegistryByScenario() {
-    final var result =
+    val result =
         service.templateFromPromptRegistryByScenario("Cloud ERP systems").getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testLocalPromptTemplate() throws IOException {
-    final var result =
+    val result =
         service
             .localPromptTemplate(
                 Files.readString(Path.of("src/main/resources/promptTemplateExample.yaml")))
             .getOriginalResponse();
-    final var choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
+    val choices = ((LLMModuleResultSynchronous) result.getOrchestrationResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
   }
 
   @Test
   void testStreamingErrorHandlingTemplate() {
-    final var template = Message.user("Bad template: {{?language!@#$}}");
-    final var templatingConfig =
+    val template = Message.user("Bad template: {{?language!@#$}}");
+    val templatingConfig =
         TemplateConfig.create().withTemplate(List.of(template.createChatMessage()));
-    final var configWithTemplate = config.withTemplateConfig(templatingConfig);
-    final var inputParams = Map.of("language", "German");
-    final var prompt = new OrchestrationPrompt(inputParams);
+    val configWithTemplate = config.withTemplateConfig(templatingConfig);
+    val inputParams = Map.of("language", "German");
+    val prompt = new OrchestrationPrompt(inputParams);
 
     assertThatThrownBy(() -> client.streamChatCompletion(prompt, configWithTemplate))
         .isInstanceOf(OrchestrationClientException.class)
@@ -380,9 +381,9 @@ class OrchestrationTest {
 
   @Test
   void testStreamingErrorHandlingInputFilter() {
-    final var prompt = new OrchestrationPrompt("Create 5 paraphrases of 'I hate you'.");
-    final var filterConfig = new AzureContentFilter().hate(AzureFilterThreshold.ALLOW_SAFE);
-    final var configWithFilter = config.withInputFiltering(filterConfig);
+    val prompt = new OrchestrationPrompt("Create 5 paraphrases of 'I hate you'.");
+    val filterConfig = new AzureContentFilter().hate(AzureFilterThreshold.ALLOW_SAFE);
+    val configWithFilter = config.withInputFiltering(filterConfig);
 
     assertThatThrownBy(() -> client.streamChatCompletion(prompt, configWithFilter))
         .isInstanceOf(OrchestrationClientException.class)
@@ -392,14 +393,32 @@ class OrchestrationTest {
 
   @Test
   void testStreamingErrorHandlingMasking() {
-    final var prompt = new OrchestrationPrompt("Some message.");
-    final var maskingConfig =
+    val prompt = new OrchestrationPrompt("Some message.");
+    val maskingConfig =
         DpiMasking.anonymization().withEntities(DPIEntities.UNKNOWN_DEFAULT_OPEN_API);
-    final var configWithMasking = config.withMaskingConfig(maskingConfig);
+    val configWithMasking = config.withMaskingConfig(maskingConfig);
 
     assertThatThrownBy(() -> client.streamChatCompletion(prompt, configWithMasking))
         .isInstanceOf(OrchestrationClientException.class)
         .hasMessageContaining("status 400 Bad Request")
         .hasMessageContaining("'type': 'sap_data_privacy_integration', 'method': 'anonymization'");
+  }
+
+  @Test
+  void testTranslation() {
+    val result = service.translation();
+    val content = result.getContent();
+    // English translated to German
+    assertThat(content).contains("Englisch");
+    assertThat(content).contains("Der", "ist");
+
+    GenericModuleResult inputTranslation =
+        result.getOriginalResponse().getModuleResults().getInputTranslation();
+    GenericModuleResult outputTranslation =
+        result.getOriginalResponse().getModuleResults().getOutputTranslation();
+    assertThat(inputTranslation).isNotNull();
+    assertThat(outputTranslation).isNotNull();
+    assertThat(inputTranslation.getMessage()).isEqualTo("Input to LLM is translated successfully.");
+    assertThat(outputTranslation.getMessage()).isEqualTo("Output Translation successful");
   }
 }